Allow customers to specify additional train/evaluation files (#4060)
Co-authored-by: Matthias Blondeel <mablonde@microsoft.com>

Evaluation tasks in MI2 only support classification
If a customer wants to finetune MI2 using multi-label data (captions), they need to be able to specify a different dataset for the evaluation portion of their data.
This change allows us to support 2 optional files in the pipeline that will be used for this evaluation.